Develop

+++ eic-registry-rest/src/main/java/eu/einfracentral/registry/controller/InfraServiceController.java
@@ -18,12 +18,12 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("infraService")
@@ -40,7 +40,8 @@ public class InfraServiceController {
     }
 
     @ApiIgnore
-    @RequestMapping(path = {"delete/{id}/", "delete/{id}/{version}/"}, method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
+//    @ApiOperation(value = "Deletes the infraService with the given id.")
+    @RequestMapping(path = {"{id}", "{id}/{version}"}, method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<InfraService> delete(@PathVariable("id") String id, @PathVariable Optional<String> version, @ApiIgnore Authentication authentication) throws ResourceNotFoundException {
         InfraService service;
@@ -48,11 +49,16 @@ public class InfraServiceController {
             service = infraService.get(id, version.get());
         else
             service = infraService.get(id);
+        if (service == null) {
+            return new ResponseEntity<>(HttpStatus.GONE);
+        }
         infraService.delete(service);
-        return new ResponseEntity<>(HttpStatus.OK);
+        logger.info("User " + authentication.getName() + " deleted InfraService " + service.getName() + " with id: " + service.getId());
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
+
     @ApiIgnore
-    @RequestMapping(path = "delete/all/", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
+    @RequestMapping(path = "delete/all", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<InfraService> deleteAll(@ApiIgnore Authentication authentication) throws ResourceNotFoundException {
         FacetFilter ff = new FacetFilter();
@@ -62,10 +68,10 @@ public class InfraServiceController {
             logger.info(String.format("Deleting service with name: %s", service.getName()));
             infraService.delete(service);
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @RequestMapping(path = {"updateFields/all/"}, method = RequestMethod.PATCH, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
+    @RequestMapping(path = {"updateFields/all"}, method = RequestMethod.PATCH, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<List<InfraService>> updateFields(InfraService service, Authentication authentication) {
         return new ResponseEntity<>(infraService.eInfraCentralUpdate(service), HttpStatus.OK);
@@ -91,20 +97,27 @@ public class InfraServiceController {
     @RequestMapping(method = RequestMethod.POST, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<InfraService> add(@RequestBody InfraService service, Authentication authentication) {
-        return new ResponseEntity<>(infraService.add(service, authentication), HttpStatus.OK);
+        ResponseEntity<InfraService> ret = new ResponseEntity<>(infraService.add(service, authentication), HttpStatus.OK);
+        logger.info("User " + authentication.getName() + " added InfraService " + service.getName() + " with id: " + service.getId() + " and version: " + service.getVersion());
+        logger.info(" Service Providers: " + service.getProviders());
+        return ret;
     }
 
     @ApiOperation(value = "Updates the InfraService assigned the given id with the given InfraService, keeping a history of revisions.")
     @RequestMapping(method = RequestMethod.PUT, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<InfraService> update(@RequestBody InfraService service, @ApiIgnore Authentication authentication) throws ResourceNotFoundException {
-        return new ResponseEntity<>(infraService.update(service, authentication), HttpStatus.OK);
+        ResponseEntity<InfraService> ret = new ResponseEntity<>(infraService.update(service, authentication), HttpStatus.OK);
+        logger.info("User " + authentication.getName() + " updated InfraService " + service.getName() + " with id: " + service.getId());
+        return ret;
     }
 
     @ApiOperation(value = "Validates the InfraService without actually changing the repository")
     @RequestMapping(path = "validate", method = RequestMethod.POST, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
     public ResponseEntity<Boolean> validate(@RequestBody InfraService service, @ApiIgnore Authentication auth) {
-        return ResponseEntity.ok(infraService.validate(service));
+        ResponseEntity<Boolean> ret = ResponseEntity.ok(infraService.validate(service));
+        logger.info("Validating InfraService " + service.getName());
+        return ret;
     }
 
     @ApiOperation(value = "Filter a list of services based on a set of filters or get a list of all services in the eInfraCentral Catalogue  ")
@@ -116,12 +129,8 @@ public class InfraServiceController {
             @ApiImplicitParam(name = "orderField", value = "Order field", dataType = "string", paramType = "query")
     })
     @RequestMapping(path = "all", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_UTF8_VALUE})
-    public ResponseEntity<Paging<InfraService>> getAll(@ApiIgnore @RequestParam Map<String, Object> allRequestParams, @ApiIgnore Authentication authentication) {
-        FacetFilter ff = new FacetFilter();
-        ff.setKeyword(allRequestParams.get("query") != null ? (String) allRequestParams.remove("query") : "");
-        ff.setFrom(allRequestParams.get("from") != null ? Integer.parseInt((String) allRequestParams.remove("from")) : 0);
-        ff.setQuantity(allRequestParams.get("quantity") != null ? Integer.parseInt((String) allRequestParams.remove("quantity")) : 10);
-        ff.setFilter(allRequestParams);
+    public ResponseEntity<Paging<InfraService>> getAll(@ApiIgnore @RequestParam MultiValueMap<String, Object> allRequestParams, @ApiIgnore Authentication authentication) {
+        FacetFilter ff = createMultiFacetFilter(allRequestParams);
         return ResponseEntity.ok(infraService.getAll(ff, authentication));
     }
 
@@ -144,7 +153,46 @@ public class InfraServiceController {
         sm.setModifiedBy("system");
         sm.setModifiedAt(String.valueOf(System.currentTimeMillis()));
         service.setServiceMetadata(sm);
+        if (active) {
+            logger.info("User " + auth.getName() + " set InfraService " + service.getName() + " with id: " + service.getId() + " to active");
+        } else {
+            logger.info("User " + auth.getName() + " set InfraService " + service.getName() + " with id: " + service.getId() + " to inactive");
+        }
         return ResponseEntity.ok(infraService.update(service, auth));
+    }
+
+    private FacetFilter createMultiFacetFilter(MultiValueMap<String, Object> allRequestParams) {
+        logger.debug("Request params: " + allRequestParams);
+        FacetFilter facetFilter = new FacetFilter();
+        facetFilter.setKeyword(allRequestParams.get("query") != null ? (String) allRequestParams.remove("query").get(0) : "");
+        facetFilter.setFrom(allRequestParams.get("from") != null ? Integer.parseInt((String) allRequestParams.remove("from").get(0)) : 0);
+        facetFilter.setQuantity(allRequestParams.get("quantity") != null ? Integer.parseInt((String) allRequestParams.remove("quantity").get(0)) : 10);
+        Map<String, Object> sort = new HashMap<>();
+        Map<String, Object> order = new HashMap<>();
+        String orderDirection = allRequestParams.get("order") != null ? (String) allRequestParams.remove("order").get(0) : "asc";
+        String orderField = allRequestParams.get("orderField") != null ? (String) allRequestParams.remove("orderField").get(0) : null;
+        if (orderField != null) {
+            order.put("order", orderDirection);
+            sort.put(orderField, order);
+            facetFilter.setOrderBy(sort);
+        }
+        if (!allRequestParams.isEmpty()) {
+            Set<Map.Entry<String, List<Object>>> filterSet = allRequestParams.entrySet();
+            for (Map.Entry<String, List<Object>> entry : filterSet) {
+                // split values separated by comma to entries and replace existing <key,value> pair with the new one
+                allRequestParams.replace(entry.getKey(), new LinkedList<>(
+                        entry.getValue()
+                                .stream()
+                                .flatMap(e -> Arrays.stream(e.toString().split(",")))
+                                .distinct()
+                                .collect(Collectors.toList()))
+                );
+            }
+            Map<String, Object> multiFilter = new HashMap<>();
+            multiFilter.put("multi-filter", allRequestParams);
+            facetFilter.setFilter(multiFilter);
+        }
+        return facetFilter;
     }
 
 }
